backend/requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Session==0.8.0
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.3.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
scikit-learn==1.7.1
scipy==1.16.1
threadpoolctl==3.6.0
Werkzeug==3.1.3
